chore: Added a Jupyter notebook
Allows me to run previous code in main python file.
This makes it easier to explore the dataset as running code in testing such as extracting a csv and conv to DF is redundant as this is not going to change after its initial run.
This therefore makes it easier to develop